fix: function location
