+++ packages/material-service/package.json
@@ -8,8 +8,7 @@
   "typings": "lib/index.d.ts",
   "dependencies": {
     "axios": "^0.19.2",
-    "lodash.kebabcase": "^4.1.1",
-    "package-json": "^6.5.0"
+    "lodash.kebabcase": "^4.1.1"
   },
   "devDependencies": {
     "@types/vscode": "^1.45.0"
