+++ packages/material-utils/package.json
@@ -5,5 +5,8 @@
     "lib"
   ],
   "main": "lib/index.js",
-  "typings": "lib/index.d.ts"
+  "typings": "lib/index.d.ts",
+  "dependencies": {
+    "package-json": "^6.5.0"
+  }
 }
