+++ tsconfig.json
@@ -7,6 +7,7 @@
     {"path": "packages/block-add"},
     {"path": "packages/iceworks-cli"},
     {"path": "packages/material-utils"},
-    {"path": "packages/material-service"}
+    {"path": "packages/material-service"},
+    {"path": "packages/page-service"}
   ]
 }
